applied PHPCsFixer hints, refactored code according to phpstan and ph…

+++ src/Service/TaxCalculator.php
@@ -11,7 +11,7 @@ class TaxCalculator
 {
     private int $taxRate = 23;
 
-    private int|float $total = 0;
+    private float|int $total = 0;
 
     private string $netAmount = '';
 
@@ -21,18 +21,18 @@ class TaxCalculator
     {
         $this->calculateTax($order->getAmount());
 
-        $order->setNetAmount((int) $this->getNetAmount());
-        $order->setVatAmount((int) $this->getVatAmount());
+        $order->setNetAmount((int)$this->getNetAmount());
+        $order->setVatAmount((int)$this->getVatAmount());
     }
 
-    public function calculateTax(int|float $total): void
+    public function calculateTax(float|int $total): void
     {
         try {
             $this->total = $total;
-            $vatDivisor = (string) (1 + ($this->taxRate / 100));
-            $this->netAmount = bcdiv((string) $this->total, $vatDivisor);
+            $vatDivisor = (string)(1 + ($this->taxRate / 100));
+            $this->netAmount = bcdiv((string)$this->total, $vatDivisor);
 
-            $this->vatAmount = bcsub((string) $this->total, $this->netAmount);
+            $this->vatAmount = bcsub((string)$this->total, $this->netAmount);
         } catch (DivisionByZeroError) {
             // no chance to throw this error, but we need to handle that for static analysis and coverage.
         }
@@ -41,7 +41,7 @@ class TaxCalculator
     public function getNetAmount(bool $humanFriendly = false): string
     {
         if ($humanFriendly) {
-            return number_format((int) $this->netAmount / 100, 2, '.', ' ');
+            return number_format((int)$this->netAmount / 100, 2, '.', ' ');
         }
 
         return $this->netAmount;
@@ -50,7 +50,7 @@ class TaxCalculator
     public function getVatAmount(bool $humanFriendly = false): string
     {
         if ($humanFriendly) {
-            return number_format((int) $this->vatAmount / 100, 2, '.', ' ');
+            return number_format((int)$this->vatAmount / 100, 2, '.', ' ');
         }
 
         return $this->vatAmount;
